Bug 1299615 - Part 3: Update in-tree ICU to release 58.1. rs=Waldo
UltraBlame original commit: eb823a4a31f084a98d08ff5974eefddb0ffafc79

+++ intl/icu/source/common/messagepattern.cpp
@@ -12,6 +12,8 @@
 
 
 
+
+
 #include "unicode/utypes.h"
 
 #if !UCONFIG_NO_FORMATTING
@@ -114,7 +116,7 @@ MessagePatternList<T, stackCapacity>::copyFrom(
             errorCode=U_MEMORY_ALLOCATION_ERROR;
             return;
         }
-        uprv_memcpy(a.getAlias(), other.a.getAlias(), length*sizeof(T));
+        uprv_memcpy(a.getAlias(), other.a.getAlias(), (size_t)length*sizeof(T));
     }
 }
