+++ intl/icu/source/common/uarrsort.c
@@ -16,6 +16,8 @@
 
 
 
+
+
 #include "unicode/utypes.h"
 #include "cmemory.h"
 #include "uarrsort.h"
@@ -120,7 +122,7 @@ doInsertionSort(char *array, int32_t length, int32_t itemSize,
         if(insertionPoint<j) {
             char *dest=array+insertionPoint*itemSize;
             uprv_memcpy(pv, item, itemSize);  
-            uprv_memmove(dest+itemSize, dest, (j-insertionPoint)*itemSize);
+            uprv_memmove(dest+itemSize, dest, (j-insertionPoint)*(size_t)itemSize);
             uprv_memcpy(dest, pv, itemSize);  
         }
     }
@@ -181,7 +183,7 @@ subQuickSort(char *array, int32_t start, int32_t limit, int32_t itemSize,
         right=limit;
 
         
-        uprv_memcpy(px, array+((start+limit)/2)*itemSize, itemSize);
+        uprv_memcpy(px, array+(size_t)((start+limit)/2)*itemSize, itemSize);
 
         do {
             while(
@@ -200,9 +202,9 @@ subQuickSort(char *array, int32_t start, int32_t limit, int32_t itemSize,
                 --right;
 
                 if(left<right) {
-                    uprv_memcpy(pw, array+left*itemSize, itemSize);
-                    uprv_memcpy(array+left*itemSize, array+right*itemSize, itemSize);
-                    uprv_memcpy(array+right*itemSize, pw, itemSize);
+                    uprv_memcpy(pw, array+(size_t)left*itemSize, itemSize);
+                    uprv_memcpy(array+(size_t)left*itemSize, array+(size_t)right*itemSize, itemSize);
+                    uprv_memcpy(array+(size_t)right*itemSize, pw, itemSize);
                 }
 
                 ++left;
