+++ intl/icu/source/common/ucharstriebuilder.cpp
@@ -12,6 +12,8 @@
 
 
 
+
+
 #include "unicode/utypes.h"
 #include "unicode/ucharstrie.h"
 #include "unicode/ucharstriebuilder.h"
@@ -115,7 +117,7 @@ UCharsTrieBuilder::add(const UnicodeString &s, int32_t value, UErrorCode &errorC
             return *this;
         }
         if(elementsLength>0) {
-            uprv_memcpy(newElements, elements, elementsLength*sizeof(UCharsTrieElement));
+            uprv_memcpy(newElements, elements, (size_t)elementsLength*sizeof(UCharsTrieElement));
         }
         delete[] elements;
         elements=newElements;
