+++ intl/icu/source/common/uniset.cpp
@@ -8,6 +8,8 @@
 
 
 
+
+
 #include "unicode/utypes.h"
 #include "unicode/parsepos.h"
 #include "unicode/symtable.h"
@@ -234,7 +236,7 @@ UnicodeSet::UnicodeSet(const UnicodeSet& o, UBool ) :
     if(list!=NULL){
         
         len = o.len;
-        uprv_memcpy(list, o.list, len*sizeof(UChar32));
+        uprv_memcpy(list, o.list, (size_t)len*sizeof(UChar32));
         if (strings != NULL && o.strings != NULL) {
             strings->assign(*o.strings, cloneUnicodeString, status);
         } else { 
@@ -286,7 +288,7 @@ UnicodeSet& UnicodeSet::operator=(const UnicodeSet& o) {
         return *this; 
     }
     len = o.len;
-    uprv_memcpy(list, o.list, len*sizeof(UChar32));
+    uprv_memcpy(list, o.list, (size_t)len*sizeof(UChar32));
     if (o.bmpSet == NULL) {
         bmpSet = NULL;
     } else {
@@ -1253,14 +1255,14 @@ UnicodeSet& UnicodeSet::complement(void) {
         if (U_FAILURE(status)) {
             return *this;
         }
-        uprv_memcpy(buffer, list + 1, (len-1)*sizeof(UChar32));
+        uprv_memcpy(buffer, list + 1, (size_t)(len-1)*sizeof(UChar32));
         --len;
     } else {
         ensureBufferCapacity(len+1, status);
         if (U_FAILURE(status)) {
             return *this;
         }
-        uprv_memcpy(buffer + 1, list, len*sizeof(UChar32));
+        uprv_memcpy(buffer + 1, list, (size_t)len*sizeof(UChar32));
         buffer[0] = UNICODESET_LOW;
         ++len;
     }
